requirements.txt
torch
torchvision
ftfy
regex
tqdm
faiss-cpu
opencv-python
Pillow
git+https://github.com/openai/CLIP.git
